requirements.txt
mysqlclient
requests==2.3.0
six==1.10.0
python-dotenv
